+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to `pasts` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://jeronlau.tk/semver/).
 
-## [0.12.0] - Unreleased
+## [0.12.0] - 2022-07-31
 ### Added
  - **`no-std`** feature